+++ src/test/org/firebirdsql/common/FBTestProperties.java
@@ -229,6 +229,8 @@ public final class FBTestProperties {
         }
         fbManager.start();
         fbManager.setForceCreate(true);
+        // disable force write for minor increase in test throughput
+        fbManager.setForceWrite(false);
         fbManager.createDatabase(getDatabasePath(), DB_USER, DB_PASSWORD);
     }